test(extension): split tab-grouping and cli tests into their own files
